// src/api/models/CreateComplaintResponseResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import {
  CreateComplaintResponseResponseBody,
  CreateComplaintResponseResponseBodyFromJSON,
  CreateComplaintResponseResponseBodyFromJSONTyped,
  CreateComplaintResponseResponseBodyToJSON,
} from './'

/**
 *
 * @export
 * @interface CreateComplaintResponseResponse
 */
export interface CreateComplaintResponseResponse {
  /**
   *
   * @type {CreateComplaintResponseResponseBody}
   * @memberof CreateComplaintResponseResponse
   */
  body: CreateComplaintResponseResponseBody | null
  /**
   *
   * @type {Array<object>}
   * @memberof CreateComplaintResponseResponse
   */
  errors: Array<object>
}

export function CreateComplaintResponseResponseFromJSON(
  json: any,
): CreateComplaintResponseResponse {
  return CreateComplaintResponseResponseFromJSONTyped(json, false)
}

export function CreateComplaintResponseResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CreateComplaintResponseResponse {
  if (json === undefined || json === null) {
    return json
  }
  return {
    body: CreateComplaintResponseResponseBodyFromJSON(json['body']),
    errors: json['errors'],
  }
}

export function CreateComplaintResponseResponseToJSON(
  value?: CreateComplaintResponseResponse | null,
): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    body: CreateComplaintResponseResponseBodyToJSON(value.body),
    errors: value.errors,
  }
}