// src/api/models/EmptyBodyResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
/**
 *
 * @export
 * @interface EmptyBodyResponse
 */
export interface EmptyBodyResponse {
  /**
   *
   * @type {object}
   * @memberof EmptyBodyResponse
   */
  body: object | null
  /**
   *
   * @type {Array<object>}
   * @memberof EmptyBodyResponse
   */
  errors: Array<object>
}

export function EmptyBodyResponseFromJSON(json: any): EmptyBodyResponse {
  return EmptyBodyResponseFromJSONTyped(json, false)
}

export function EmptyBodyResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): EmptyBodyResponse {
  if (json === undefined || json === null) {
    return json
  }
  return {
    body: json['body'],
    errors: json['errors'],
  }
}

export function EmptyBodyResponseToJSON(value?: EmptyBodyResponse | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    body: value.body,
    errors: value.errors,
  }
}